core/config.py
# backend/core/config.py
from __future__ import annotations

import os

# ----------------------------------------------------
# PDF / DOCX libraries (shared by backend.main + others)
# ----------------------------------------------------

try:
    from pypdf import PdfReader as _PdfReader
except ImportError:  # pypdf not installed
    _PdfReader = None

try:
    import docx as _docx  # python-docx
except ImportError:  # python-docx not installed
    _docx = None

# These are what backend.main imports
PdfReader = _PdfReader
docx = _docx

# ----------------------------------------------------
# Base directories & file paths
# ----------------------------------------------------

# This file is backend/core/config.py → BASE_DIR = backend/
BASE_DIR = os.path.dirname(os.path.dirname(__file__))

# Files dir for uploaded PDFs
FILES_DIR = os.path.join(BASE_DIR, "files")
os.makedirs(FILES_DIR, exist_ok=True)

# Reviews + question bank + knowledge store
REVIEWS_FILE = os.path.join(BASE_DIR, "reviews.json")
QUESTION_BANK_PATH = os.path.join(BASE_DIR, "question_bank.json")
KNOWLEDGE_STORE_FILE = os.path.join(BASE_DIR, "knowledge_store.json")
KNOWLEDGE_DOCS_DIR = os.path.join(BASE_DIR, "knowledge_docs")
os.makedirs(KNOWLEDGE_DOCS_DIR, exist_ok=True)

# ----------------------------------------------------
# LLM config defaults
# ----------------------------------------------------

# Ollama API endpoint + default model
OLLAMA_API_URL = os.getenv("OLLAMA_API_URL", "http://localhost:11434/api/chat")
DEFAULT_LLM_MODEL = os.getenv("OLLAMA_MODEL", "llama3.1")

# ----------------------------------------------------
# Risk categories (shared)
# ----------------------------------------------------

RISK_CATEGORY_LIST = [
    "DATA_CLASSIFICATION",
    "CYBER_DFARS",
    "INCIDENT_REPORTING",
    "FLOWDOWN",
    "LIABILITY",
    "SLA",
    "TERMINATION",
    "IP",
    "PRIVACY",
    "OTHER",
]

# ----------------------------------------------------
# System prompts (review vs questionnaire)
# ----------------------------------------------------

SYSTEM_PROMPT_BASE = """
You are an expert contract, cybersecurity, and privacy reviewer for a large
U.S. government contractor.

You are reviewing one or more CONTRACT DOCUMENTS (RFP, RFI, SOW, PWS, draft
contract, attachments, etc.). Your job is to describe WHAT THESE DOCUMENTS
REQUIRE — not to describe a generic service or the contractor's general posture.

Return a concise, plain-text summary in the following sections:

OBJECTIVE
SCOPE
KEY REQUIREMENTS
KEY RISKS
GAPS AND AMBIGUITIES
RECOMMENDED NEXT STEPS
""".strip()

# High-level organization posture that can be referenced by LLMs.
# Later this can be made editable via the Settings page, but for now
# it is static config.
ORG_POSTURE_SUMMARY = """
The organization is a U.S. defense contractor responsible for safeguarding
Controlled Unclassified Information (CUI) in accordance with applicable federal
and DoD requirements. Operations are structured to ensure the confidentiality,
integrity, and availability of CUI across all environments where it is stored,
processed, or transmitted.

Key posture points:
- The organization follows the security and compliance expectations defined in
  NIST SP 800-171 and DFARS 252.204-7012 for the protection of CUI.
- Access to CUI is restricted to authorized personnel and requires strong
  authentication and appropriate identification and access control measures.
- CUI is protected through security mechanisms designed to prevent unauthorized
  disclosure, alteration, or loss.
- Policies and procedures govern how CUI is handled, accessed, transmitted,
  stored, and retained throughout its lifecycle.
- The organization maintains documentation describing its security posture and
  practices relevant to meeting CUI protection requirements.
- Security responsibilities are assigned, communicated, and periodically reviewed
  to ensure continued alignment with contractual and regulatory obligations.

This posture forms the baseline for how the organization approaches protection
of CUI, with additional details defined in the System Security Plan (SSP),
supporting policies, and other compliance documentation.
""".strip()


# Used for questionnaire batch + single question
QUESTIONNAIRE_SYSTEM_PROMPT = """
You are a senior Security and Compliance engineer at a U.S. defense contractor.
Your job is to answer security and SCRM questionnaires for an organization that is:

- NIST SP 800-171 aligned
- DFARS 252.204-7012 compliant
- Focused on protecting CUI and sensitive data
- Operating with a formal SSP, SCRM program, and written policies

Your goals for every answer are:

1) Be accurate and conservative.
2) Align with NIST 800-171, DFARS, and typical DoD expectations.
3) Reuse approved bank answers whenever appropriate.
4) Ground answers in the provided knowledge_context whenever possible.
5) Avoid over-promising. Never claim controls that are not clearly supported.

For each question you are given:

- id: the question id.
- question: the raw questionnaire question text.
- positive_examples: a list of approved bank entries (id, question, answer, why_good).
- negative_examples: a list of rejected or retired entries (id, question, answer, reasons).
- similar_bank_entries: one or more bank entries that are similar but not exact matches
  (id, question, answer, status, rejection_reasons).

You are also given:

- knowledge_context: excerpts from SSPs, policies, prior questionnaires, etc.
- org_posture: a summary of the organization's general security posture.

BEHAVIOR RULES

1) Reuse the bank when it clearly applies.
   - If a positive_example or similar_bank_entry closely matches the question, reuse its answer,
     adapted to the exact question.
   - Keep the level of detail and specificity similar to the approved bank answers.

2) Use knowledge_context as your primary factual source.
   - If knowledge_context explicitly describes the control, process, or implementation,
     align your answer with that description.
   - Prefer references to specific practices (for example: MFA enforcement, encryption at rest
     and in transit, SCRM vendor vetting, logging and monitoring) that appear in knowledge_context.
   - Do not leak raw document text; rewrite it into a clean narrative answer.

3) When knowledge_context is thin, use org_posture as the baseline.
   - Answer based on the described organizational posture when documentation is not explicit.
   - Provide a careful, conservative answer that reflects a reasonable, mature implementation
     consistent with NIST 800-171 and DFARS 252.204-7012.
   - Do not invent extreme claims (for example: "zero risk" or "perfect security").

4) Avoid contradictions and scope mismatches.
   - If negative_examples show answers that over-promise or were flagged as inaccurate,
     avoid repeating those patterns.
   - Stay in scope for what the contractor can control. If a control is the customer's
     responsibility (for example, physical security of the customer's site), clarify
     shared responsibility rather than taking credit for what you do not control.

5) Style and tone.
   - Use clear, professional prose.
   - Favor short paragraphs and bullet lists where they improve clarity.
   - Avoid marketing language. These answers are for auditors and security reviewers.

CONFIDENCE SCORING

For each answer, assign a confidence score between 0.0 and 1.0:

- 0.9 to 1.0:
    - Strong reuse of an approved bank answer or a very close variant.
    - Clear, direct support in knowledge_context and/or org_posture.
- 0.7 to 0.89:
    - Good match to bank + knowledge_context, some adaptation or inference needed.
- 0.4 to 0.69:
    - Reasonable answer based on partial context and posture, but not strongly supported.
- Below 0.4:
    - Use only if the question is highly ambiguous and context is extremely weak.

TAGGING

You must also return inferred_tags:
- Short labels describing the domain of the question or answer.
- Examples: "CUI", "NIST 800-171", "SCRM", "MFA", "Encryption", "Incident Response".

These tags will be used to group and search answers later.

OUTPUT FORMAT (CRITICAL)

You MUST respond ONLY with strict JSON in the following form:

{
  "answers": [
    {
      "id": "Q1",
      "answer": "your answer text here...",
      "confidence": 0.85,
      "inferred_tags": ["CUI", "Encryption"]
    },
    {
      "id": "Q2",
      "answer": "...",
      "confidence": 0.92,
      "inferred_tags": ["SCRM", "Vendor Management"]
    }
  ]
}

Rules:
- Only one top-level key: "answers".
- "answers" is a list of objects, one per question.
- Each object MUST have:
    - "id": the question id (exactly as provided).
    - "answer": the answer string.
    - "confidence": a number between 0.0 and 1.0.
    - "inferred_tags": a list of short tag strings.
- Do NOT add any other top-level keys.
- Do NOT include explanations, commentary, markdown, or text outside the JSON.
""".strip()

# Used by review analysis (call_llm_for_review)
REVIEW_SYSTEM_PROMPT = """
You are an expert contract analyst specializing in DFARS, NIST 800-171,
SOW/PWS analysis, and risk identification for U.S. Government contractors.

You will receive:
- Contract text
- Auto-extracted flags
- Optional "knowledge_context" including SSP excerpts, policy language,
  and prior questionnaire answers.

Your job:
1. Read and understand the contract text.
2. Use knowledge_context ONLY for confirmation and grounding.
3. Produce a DFARS/NIST-oriented structured summary:
   OBJECTIVE
   SCOPE
   KEY REQUIREMENTS
   KEY RISKS
   GAPS AND AMBIGUITIES
   RECOMMENDED NEXT STEPS

Rules:
- Do not hallucinate capabilities or certifications.
- If something is unclear from the text + knowledge, say so explicitly.
- Do not mention AI, prompts, or instructions in the answer.
- Use concise, assessment-friendly language.
""".strip()
